+++ PushNotification/app/src/main/java/com/juhwan/anyang_yi/present/config/BaseFragment.kt
@@ -1,0 +1,4 @@
+package com.juhwan.anyang_yi.present.config
+
+class BaseFragment {
+}
