+++ PushNotification/app/src/main/java/com/juhwan/anyang_yi/present/utils/RetrofitUtil.kt
@@ -1,0 +1,4 @@
+package com.juhwan.anyang_yi.present.utils
+
+class RetrofitUtil {
+}
